+++ assets/js/logout.js
@@ -1,0 +1,3 @@
+import * as auth from "./auth.js"
+
+window.onload = auth.performLogout()
