+++ pages/login.md
@@ -4,20 +4,5 @@ title: TPEN
 permalink: /login/
 modules: [/assets/js/login.js]
 ---
-> `This site is scheduled for launch in 2026 and not intended for public use at this time.`{: .notice}
-
-## Transcription for Paleographical and Editorial Notation
-
----
-
-<div>
-  TPEN3 Login.  You will be redirected.
-</div>
-
----
-
-![old TPEN logo]({{site.baseurl}}/assets/img/tpen_clearShadowSmall.png)
-This tool is the evolution of the 21st century transcription platform at [t‑pen.org](https://t-pen.org).
-Originally targeting Medieval manuscripts, users have fostered a community of carefully
-transcribed images of scripts, engravings, and calligraphy from all over the world across time.
+You will be logged in and redirected
