+++ pages/logout.md
@@ -1,0 +1,7 @@
+---
+layout: module
+title: TPEN
+permalink: /logout/
+modules: [/assets/js/logout.js]
+---
+You will be logged out and redirected
